Measuring playing_position behaviour
Set up a 2-stage bucket delay to calculate the difference between
successive observed playing_positions. Implement triggers to bang on
different delay values and report playing_position when difference ==0
or <0 or >0.5.

Observations show that playing_position doesn't update for 2-3 cycles
(==0) when changing loop_end. After that we can sometimes see a big jump
(>0.5) when doubling the loop length. It seems to be exactly half the
new loop length.